lab2_postgres/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy[asyncio]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
